refactor: convert to classes
