+++ src/conditions/util/titleMatches.ts
@@ -1,5 +1,5 @@
-import { IssueProps, PRProps } from '.'
-import { processRegExpPattern } from '../utils'
+import { IssueProps, PRProps } from '../'
+import { utils } from '../../utils'
 
 const TYPE = 'titleMatches'
 
@@ -12,7 +12,7 @@ const titleMatches = (
   condition: ConditionTitleMatches,
   issue: IssueProps | PRProps
 ) => {
-  const pattern = processRegExpPattern(condition.pattern)
+  const pattern = utils.processRegExpPattern(condition.pattern)
   return pattern.test(issue.title)
 }
