+++ src/types.d.ts
@@ -28,6 +28,43 @@ export interface Config {
   delete_labels: boolean
 }
 
+interface Props {
+  creator: string
+  description: string
+  locked: boolean
+  state: 'open' | 'closed'
+  title: string
+}
+
+export interface PRProps extends Props {
+  branch: string
+  isDraft: boolean
+  files: string[]
+}
+
+export interface IssueProps extends Props {}
+
+export interface Label {
+  name: string
+  description: string
+  color: string
+}
+
+export type Labels = Label[]
+
+interface GeneralContext {
+  labels: Labels
+  IDNumber: number
+}
+
+export interface PRContext extends GeneralContext {
+  prProps: PRProps
+}
+
+export interface IssueContext extends GeneralContext {
+  issueProps: IssueProps
+}
+
 export interface Options {
   configPath: string
   showLogs: boolean
